Updated version information.
--HG--
branch : multi-monitor-support
extra : convert_revision : svn%3A4bd69a10-c447-994a-b4ab-e5eff1cba757/branches/multi-monitor-support%40138

+++ Windawesome/Window.cs
@@ -172,7 +172,7 @@ namespace Windawesome
 
 			// show/hide from Taskbar
 			NativeMethods.PostMessage(Windawesome.taskbarButtonsWindowHandle, NativeMethods.WM_SHELLHOOKMESSAGE,
-				(UIntPtr) (show ? NativeMethods.ShellEvents.HSHELL_WINDOWCREATED : NativeMethods.ShellEvents.HSHELL_WINDOWDESTROYED), hWnd);
+				(UIntPtr) (uint) (show ? NativeMethods.ShellEvents.HSHELL_WINDOWCREATED : NativeMethods.ShellEvents.HSHELL_WINDOWDESTROYED), hWnd);
 		}
 
 		internal void ToggleShowHideTitlebar()
